Service levels filtering (#65)
* Add Service Levels filtering tab
(re: 46)

This includes filtering properties for :
- Avg. Frequency ≦
- Avg. Frequency <
- Calculate frequency based on single routes
- Maximum fare
- Minimum fare
(fare filtering is disabled for now)

* Restore the unusual filtering rules to support `selectedDays`
re: #60

Note, `selectedDays` is special..
Unlike the other query parameters, we want to allow it to be an empty string ''.
The empty string means that the user unchecked all the days.
We don't want to clear out the parameter and re-check them if the user does this.

* Add min and step attributes to the numeric frequency and fare inputs

`min=0` prevents these fields from going negative
`step=0.01` allows 2 decimal places in the fare inputs
Also made the fare inputs slightly wider to accomodate the 2 decimals

* pacify eslint

* Fix merge

---------

Co-authored-by: Ian Rees <ian@ianrees.net>