Add Filter Summary to report component
This generates a filter summary as a computed property to add
plain-text descriptions of the established filters.
For example, "* operating between 03/17/2025 and 03/24/2025"

This also adjusts the date/time formatter functions to accept an
optional format argument that overrides the default format string.
It is used to display the dates and times in friendlier format.