Don't show legend details when there is no data / bbox detail (closes #83)
This adds a computed property `hasData` to track whether there
is data to show on the map and in the legend.
We use this to hide the legend's style info if there is no data.

This also handles the situation where the bounding box is gone
and the query is running but the data hasn't loaded yet.
In this situation we show a spinner in the legend div while we
wait for the style data to be ready.