Sprint 3 polish items (#98)
* Don't show legend details when there is no data / bbox detail
(closes #83)

This adds a computed property `hasData` to track whether there
is data to show on the map and in the legend.
We use this to hide the legend's style info if there is no data.

This also handles the situation where the bounding box is gone
and the query is running but the data hasn't loaded yet.
In this situation we show a spinner in the legend div while we
wait for the style data to be ready.

* Swap CALACT logo for ODOT logo on home page
(closes #80)

* Disable the download button while data is still loading
(closes #90)

This passes `stopDepartureLoadingComplete` to the report component
to disable the button.

This commit also changes `stopDepartureLoadingComplete` from a
two-way model bind to just a normal one-way v-bind in the filter
component - the property should only change up at tne, so two
way binding is unnecessary.

* Adjust z-index so share button and menu go above legend
(closes #97)

* Add toggle to maximize/minimize the legend
(closes #96)

This uses the Oruga collapse control

* Switch to 24 hr time pickers (removes am/pm)
(closes #95)

This also fixes some lint warnings in this file.

* Reset state when running additional queries
(closes #93)

This replaces `setReady()` and the `ready` boolean to be `runQuery()` and `runCount` number.
This change makes it easier to increment the run counter as the user runs more reports.

* stopDepartureLoadingComplete in map-share component

---------

Co-authored-by: Ian Rees <ian@ianrees.net>