Scenario streaming (#161)
Refactor scenario fetcher to use a streaming approach.

Essentially, it emits "ScenarioProgress" events that contain partial data updates. A listener can receive these callbacks and build up the complete ScenarioData object from these partial updates.

Adapters are provided for "classic" behavior, where a "ScenarioDataReceiver" is directly connected to "ScenarioFetcher" and after all processing is done, returns a "ScenarioData" object that is fully loaded and indexed.

Additionally, a "streaming" adapter is provided that places a ScenarioStreamSender/ScenarioStreamClient in the middle of this process and transports the data over a network or file stream.